requirements.txt
# Trading AI System Dependencies

# AI APIs
anthropic>=0.18.0
google-generativeai>=0.4.0

# Market Data & Broker
alpaca-trade-api>=3.0.0
yfinance>=0.2.36

# Technical Analysis
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
mplfinance>=0.12.10b0

# News & Sentiment Analysis
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
textblob>=0.17.1
newspaper3k>=0.2.8

# Web scraping & async
aiohttp>=3.9.0
asyncio>=3.4.3

# Scheduling & Time
schedule>=1.2.1
pytz>=2024.1

# Data Storage
sqlalchemy>=2.0.0
sqlite3

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging & Monitoring
rich>=13.7.0
loguru>=0.7.2

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0
